CB-32487 Experimental forced reboot for RHEL9 ARM64

+++ scripts/reboot.sh
@@ -1,4 +1,9 @@
 #!/bin/bash
 
+# This standard reboot command causes problems with RHEL 9 ARM64 images and right now, not sure why...
 echo "Rebooting before validation and cleanup"
-reboot
+if [[ "${OS}" != "redhat9" || "${ARCHITECTURE}" != "arm64" ]] ; then
+  reboot
+else
+  sudo systemctl reboot --force
+fi
